Added batch mean collector

+++ src/main/java/it/simulation/system/SystemStats.java
@@ -10,10 +10,10 @@ public class SystemStats {
     private final double meanBusyServer;      // Mean busy server over time
     private final double meanServiceTime;     // Mean system service time
     private final double meanResponseTime;    // Mean system response time
-    private final double totalCompletion;     // Sum of completion over all web server
+    private final int totalCompletion;     // Sum of completion over all web server
     private final double totalBusyTime;       // Sum of busy time over all web server
 
-    public SystemStats(double throughput, double meanUtilization, double meanBusyServer, double meanServiceTime, double meanResponseTime, double totalCompletion, double totalBusyTime) {
+    public SystemStats(double throughput, double meanUtilization, double meanBusyServer, double meanServiceTime, double meanResponseTime, int totalCompletion, double totalBusyTime) {
         this.throughput = throughput;
         this.meanUtilization = meanUtilization;
         this.meanBusyServer = meanBusyServer;
